refactor: Reorganizes aws package
The `aws` package use by the `legalbrawlapi` module is getting
needlessly deep, and is causing circular dependencies. Because the API
gateway request body is being used by DynamoDB in `PutItem` requests
(among others).

This "flattening" of the structure should better support the
requirements that these two components have.
