+++ backend/src/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"legalbrawlapi/aws/apigateway"
+	lbapiaws "legalbrawlapi/aws"
 
 	"github.com/aws/aws-lambda-go/lambda"
 )
 
 func main() {
-	lambda.Start(apigateway.HandleRequest)
+	lambda.Start(lbapiaws.HandleRequest)
 }
